backend/python-services/requirements.txt
# backend/python-services/requirements.txt

Flask
pytesseract
Pillow
pdf2image
numpy
opencv-python
gunicorn